apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.19
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
openai==1.57.4
anthropic==0.40.0
APScheduler==3.10.4
pydantic==2.10.3
pydantic-settings==2.7.0